Change (HL) jump operands to RegisterPairImplied
While the usual notation implies that these operands
are MemoryIndirect, since no memory is accessed during
these instructions, implementations relying on the
operand type to determine what data to access would
have to treat jump instructions as a special case.

This requires adding extra branches in the Display
implementation for Instruction, and could lead to
confusion for users, so unfortunately there are
tradeoffs involved with this decision.