Return Inva instructions for ignored index prefixes
This is similar to the old Noni instruction type,
but this time we correctly return Inva for all
invalid indexed instructions, and not just for
double prefixes. An indexed Inva instruction
represents a single wasted M1 machine cycle.

The returned instruction has an opcode value
matching the second byte in the decoded sequence,
but it is redundant and should not be considered
part of the decoded instruction, as the following
decoding cycle begins with that opcode.

This reintroduces the Peekable dependency, but
the previous implementation would consume both
bytes of the invalid instruction, which is
incorrect anyway.