Add OctetImplied operand type
The unofficial instruction OUT (C),0 is
a special case in that it has an implied
operand value, rather than an explicit one.

Instead of handling it by checking the
instruction type and the operand combination,
this adds an operand type that correctly
describes it, and makes it easier to detect
and handle both in DeZ80 (e.g. by avoiding
it outputting a 0-valued byte in
Instruction::to_bytes()), and in software
using DeZ80 for instruction decoding.