Separate push and decode calls in InstructionDecoder
While adding one line to the decoding loop in some
cases, it is generally more flexible in that bytes
can be added to the decoder state and accumulated
prior to any decoding attempt.

This adds push_byte() and push_slice() functions
to modify the decoder state, as well as a
try_decode() function for decoding. The
decode_byte() and decode_slice() functions are
deprecated.

try_decode() returns a
Result<Instruction, PartialDecodeResult> instead of
an Option<Instruction> as in the deprecated API
and the stateless decoding functions. This allows
callers to get some information on instructions
being decoded before they can be fully constructed.