Update module visibility
- Exposes InstructionDecoder, DecodingState, and Instruction
at the top level of the crate
- Removes top-level visibility for register data structures
- Removes public visibility for the decoder module

The goal is to make imports self-documenting by encouraging verbosity,
while keeping the most important structures accessible at the top level.
For example, it is not obvious that a dez80::Condition relates to
a conditional instruction rather than something related to the decoder,
whereas importing dez80::Instruction::Condition makes the association
clearer.