Add DecodingState variants for displacement operands
Displacement operands affect behavior significantly,
so in a progressive decoding scenario it is useful to
distinguish between displacement and non-displacement
operands.

Without distinct DecodingState variants or another
signaling mechanism, it is not always possible to detect
a displacement operand with limited context of the
decoder's state, because indexed instruction tables
include some instructions with only non-displacement
operands, as well as both types of operands.

While disambiguation is only strictly necessary for
indexed instructions, for consistency this also replaces
the IndexedBitwiseOperand variant with
IndexedBitwiseDisplacement, since all bitwise indexed
instructions have a displacement as their only operand,
as well as RootDisplacement, used in relative jumps.